refactor: move DepsT into context package
