+++ python/mirascope/llm/calls/context_call.py
@@ -4,12 +4,12 @@ from collections.abc import Sequence
 from dataclasses import dataclass
 
 from ..content import UserContent
-from ..context import Context
+from ..context import Context, DepsT
 from ..prompts import Prompt
 from ..response_formatting import FormatT
 from ..responses import Response
 from ..streams import AsyncStream, Stream
-from ..types import DepsT, P
+from ..types import P
 from .base_context_call import BaseContextCall
