+++ python/mirascope/llm/clients/openai.py
@@ -6,7 +6,7 @@ from typing import Any, TypeAlias
 from openai.types.chat import ChatCompletionMessageParam
 
 from ..clients import BaseParams
-from ..context import Context
+from ..context import Context, DepsT
 from ..messages import Message
 from ..response_formatting import FormatT
 from ..responses import Response
@@ -15,7 +15,6 @@ from ..streams import (
     Stream,
 )
 from ..tools import ContextToolDef, ToolDef
-from ..types import DepsT
 from .base import BaseClient
 from .register import OPENAI_REGISTERED_LLMS
