+++ python/mirascope/llm/context/__init__.py
@@ -1,5 +1,5 @@
 """The Context module for providing context to LLM calls during generation."""
 
-from .context import Context, context
+from .context import Context, DepsT, context
 
-__all__ = ["Context", "context"]
+__all__ = ["Context", "DepsT", "context"]
